+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - General
   - Add support for LSP method `textDocument/prepareRename` which it's the proper way to check if the rename will work correctly. #642 
   - Expose new custom method `clojure/cursorInfo/raw` for custom hack on current cursor information code. #645
+  - Support stub generation using `clj-easy/stub`, adding analysis and liniting support for closed sources codes like Datomic. Check `:stubs` settings for more details. #637
 
 ## 2021.11.16-16.52.14